/login endpoint fully functional with token json return

+++ backendAPI/routes/api.php
@@ -13,14 +13,15 @@ use Illuminate\Support\Facades\Route;
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
+Route::put('user/login', 'UserController@userLogin'); //login doesnt go trough auth guard
+Route::post('/user', 'UserController@store')->name('user.store');
 
-Route::group(['middleware' => 'auth:api', 'prefix' => 'user'], function() {
-    Route::get('/', 'UserController@index');
-    Route::post('/', 'UserController@store');
-    Route::put('/{id}', 'UserController@update');
-    Route::delete('/{id}', 'UserController@destroy');
-    Route::get('/search', 'UserController@search');
-    Route::put('/login', 'UserController@userLogin');
+Route::group(['prefix' => 'user'], function() {
+    Route::get('/', 'UserController@index')->middleware('auth:api')->name('user.index');
+    Route::put('/{id}', 'UserController@update')->middleware('auth:api')->name('user.update');
+    Route::delete('/{id}', 'UserController@destroy')->middleware('auth:api')->name('user.destroy');
+    Route::get('/search', 'UserController@search')->middleware('auth:api')->name('user.search');
+
 });
 
 Route::apiResource( 'tickets', 'TicketsController' );
